Organize directory structure of builds repository
Organize directory structure of builds repository
