+++ config/host_os.yaml
@@ -24,7 +24,7 @@ build_packages:
  force_rebuild: False
  mock_config:
   CentOS:
-   '7': "./mock_configs/CentOS/7/CentOS-7-ppc64le.cfg"
+   '7': "./config/mock/CentOS/7/CentOS-7-ppc64le.cfg"
  mock_args:
    # If the system has less RAM than required, the plugin will be disabled
    "--enable-plugin=tmpfs
@@ -42,7 +42,7 @@ build_iso:
     iso_name: "OpenPOWER-Host_OS"
     mock_config:
      CentOS:
-      '7': "./mock_configs/CentOS/7/build-iso-CentOS-7-ppc64le.cfg"
+      '7': "./config/mock/CentOS/7/build-iso-CentOS-7-ppc64le.cfg"
     mock_args:
       # If the system has less RAM than required, the plugin will be disabled
       "--enable-plugin=tmpfs
