+++ config/mock/CentOS/7/CentOS-7-ppc64le.cfg
