+++ config/mock/README.rst
