+++ lib/config.py
@@ -214,7 +214,7 @@ class ConfigParser(object):
                                  help='Path of the configuration file for build '
                                       'scripts',
                                  # NOTE(maurosr): move this to /etc in the future
-                                 default='./config.yaml')
+                                 default='config/host_os.yaml')
         self.parser.add_argument('--verbose', '-v',
                                  help='Set the scripts to be verbose',
                                  action='store_true')
